examples/firered-image-edit/pyproject.toml
[project]
name = "firered-image-edit"
version = "0.1.0"
description = "FireRed Image Edit example (inference-only; Cozy manifest via cozy.toml)"
requires-python = ">=3.12"
dependencies = [
    "gen-worker[torch]==0.3.0",
    "diffusers @ git+https://github.com/huggingface/diffusers",
    "transformers>=4.51.3",
    "accelerate",
    "pillow",
    "sentencepiece",
    "protobuf",
]

[dependency-groups]
dev = ["mypy>=1.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/firered_image_edit"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true

